new: add divider
This utility class implements unsigned 64-bit division by a constant as
multiplication, based on the information in

	* Hacker's Delight, 2nd edition §10 (Warren, 2003) [1]

and the algorithms described in

	* Division by Invariant Integers using Multiplication §8 (Granlund &
	  Montgomery, 1994) [2],
	* Improved division by invariant integers (Möller & Granlund, 2011) [3], &
	* Labor of Division, Episode III: Faster Unsigned Division by
	  Constants (ridiculous_fish, 2011) [4].

In order to facilitate several optimized versions of the division (by one,
by a power of two, without overflow), function pointers are used to select
the specific algorithm for the chosen denominator.  In order to avoid
storing the shift amount, it is baked into the method via a template
argument and picked by a switch based on the calculated shift.

This class also implements a modulo method; since it does not store the
original denominator, this must be passed by the user.  The intention is to
use this with lite::table since the number of buckets rarely changes.

My testing indicated that divider.modulo() takes roughly one-third the time
that operator% takes for uint64_ts on my 64-bit x86.  YMMV.

[1] http://www.hackersdelight.org/
[2] https://gmplib.org/~tege/divcnst-pldi94.pdf
[3] https://gmplib.org/~tege/division-paper.pdf
[4] http://ridiculousfish.com/files/faster_unsigned_division_by_constants.pdf