random: replace jrand48/erand48 with MSWS RNG
Rather than using the 48-bit libc PRNGs, swap in the Middle Square Weyl
Sequence RNG [1], which have 64-bit seeds and 128-bits of state.  Like the
old generator, this will still be initalized with system entropy by default.

Note: when generating 64-bit values, it is not clear to me if concatenating
sequential 32-bit outputs from the generator is kosher.  If not, it will
obviously have to change.

[1] https://arxiv.org/pdf/1704.00358v5.pdf