random: fix up u64 seed leakage
Previously, we were returning the entire seed, not just the truncated
result.  This could lead to leakage of the upper half of the seed if used
improperly.  It turns out that u64 was using it improperly, and in fact only
setting the high bits of its own result.  This commit modifies the base
generator to only return 32-bits and additionally fixes the 64-bit math.