test: add naïve statistics to randomness test
We use Welford's online algorithm to calculate mean & variance for a perfect
coin vs. randomly generating bools, unsigned longs, floats, and doubles.
We also generate a student's t test value that bits are generated in equal
proportions.  These tests are very basic and may not be implemented
correctly, but do seem to show that the generator is working as expected.