random: add gaussian generator(s) for floats (and doubles)
This commit adds gaussian generators for both floats and doubles based on
the Marsaglia-Bray polar method and it appears to generate with the correct
mean and standard deviation assuming that the Welford statistics are being
calculated correctly.

One unfortunate thing about this code is that the double generator `gg` is
mostly copy-pasta from the float generator `g`.  In the future, it probably
makes sense to make this trait based.

One nice thing about this code, is that it is avoids IEEE754 as long as
possible.  Rather than converting uniforms directly to floating-point to
determine whether the generated coördinate is outside a unit circle, it
instead calculates the perimeter of the unit circle in terms of the uniform
type.  Once the coördinates are accepted, it generates the floating-point
values and completes the calculation.  Another nice thing is that instead of
generating a value in [0,UMAX], converting to [0.0,1.0], and then shifting
to [-1.0,1.0]; it instead utilizes the fact that the uniform generator makes
signed integers and divides by the signed maximum instead.  It might be
worth someday removing the additional negative signed at the end of the
number line, but meh for now.

Both the float and double generators rely on 32-bit signed uniforms.  The
double generator could likely benefit from 64-bit signed uniforms, but the
current copy-pasta implementation would directly translate into something
that requires on 128-bit signeds/unsigneds.  While the status quo is leaving
loads of possible values on the table, the added complexity and additional
uniform generator calls seem like they're not worth it.

Both generators are able to calculate two independent gaussian results, but
one is thrown away.  A future improvement might save/reply one, but this
would require extra state in the uniform generator, even if gaussians were
never called.  So again, meh.

It might be worth revisiting Box-Muller in the future but Ziggurat is more
complicated than I'm interested implementing right now.