random: make sure minimum is non-zero
Prior to this change, the -n%n expression is being optimized to zero, which
is unfortunate because it means that the rejection loop is not throwing out
non-uniform values correctly.  This becomes especially evident when working
with 8-bit values.  The fastpath shows a much lower t-test values than the
slowpath.  After this change, there is no real difference.

Q.v. https://lemire.me/blog/2020/10/28/what-the-heck-is-the-value-of-n-n-in-programming-languages/