objectstack: get rid of superflous pointer
Previously, we were storing a pointer to the object to be destroyed, but
this is unnecessary.  Rather than passing that to objectstack::destroy<T>,
pass a linkage* and reinterpret it as the appropriate wrapper<T>* inside of
destroy<T>.  This lets C++ do all the messy struct math there instead of in
objectstack::make & objectstack::unwind.