haystack: refactor quote method
The refactor does two significant things.  First, it scans for spans of 's &
non-'s, precalculating the amount of space needed before reserving room in
the haystack.  Second, it coalesces 's rather than individually quoting
them.  This changes the test vectors, but the results should be the same in
any shell.