haystack: fix up alignment checks, add helpers
The previous alignment checks were totally wrong, since they were comparing
alignof (in bytes) with alignment_mask (a bitmask).  This commit fixes the
checks by adding a helper that converts alignof to a mask and a templated
predicate that can be used to verify whether the current alignment is good
enough for a partiular type.  Additionally, adds a method that aligns to a
particular type and fixes up const-correctness for several methods.