new: objectpool
Like objectstack, uses a haystack to allocate/initialize objects.  Unlike
objectstack, all the objects are of the same type/sizeof.  When an object is
deallocated, add the storage to a freelist instead of expanding the
haystack.  If the type has a non-trivial destructor, also keep a list of
objects that are live and need to be cleaned up later.  When the objectpool
is out of scope, do this cleanup.

An early implmentation attempted to use std::conditional/std::enable_if to
remove superfluous list/link from wrapper in case the destructor is trivial,
but ran into C++11 defect re: explicit specialization in non-namespace
scope, which is fixed in C++17.  Q.v. CWG 727.

Can explicitly opt-out of this autodestruct behavior with second template
argument, e.g. if caller assures that all objects will be destroyed prior
to the objectpool going out of scope.  YMMV.

If autodestruct is off, then we don't need to keep the full list of objects
or traverse it during clear.  Consequently, we don't ever need to know
whether a given object is in the free list vs. the live list, and can
therefore safely share the linkage with object storage.